Rebundle at_app with the new bundler
