--- packages/at_app/lib/src/bundles/demos/chit_chat/chit_chat_template_bundle.dart
@@ -1,8 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-import 'package:at_app_create/at_app_create.dart';
-import 'chit_chat_bundle.dart';
-
-class ChitChatTemplateBundle extends AtTemplateBundle<AtTemplateVars> {
-  ChitChatTemplateBundle() : super(chitChatBundle);
-}
