--- packages/at_app/lib/src/bundles/demos/snackbar/snackbar_template_bundle.dart
@@ -1,8 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-import 'package:at_app_create/at_app_create.dart';
-import 'snackbar_bundle.dart';
-
-class SnackbarTemplateBundle extends AtTemplateBundle<AtTemplateVars> {
-  SnackbarTemplateBundle() : super(snackbarBundle);
-}
