+++ packages/at_app/lib/src/bundles/demos/snackbar_sender/snackbar_sender.dart
@@ -1,32 +1,39 @@
-// GENERATED CODE - MODIFY AS NECESSARY
-// PREVENT OVERWRITING THIS FILE BELOW:
-// LOCKED = FALSE
+// GENERATED CODE - DO NOT MODIFY BY HAND
 
 import 'package:at_app_create/at_app_create.dart';
 import 'package:at_app/src/models/at_app_template.dart';
 import 'package:pub_semver/pub_semver.dart';
 
-import 'snackbar_sender_template_bundle.dart';
+import 'snackbar_sender_bundle.dart';
 
-export 'snackbar_sender_template_bundle.dart';
+class SnackbarSenderTemplateBundle extends AtTemplateBundle<AtTemplateVars> {
+  SnackbarSenderTemplateBundle() : super(snackbarSenderBundle);
+}
 
-final snackbarSenderTemplate = AtAppTemplate(
+final AtTemplateVars _vars = AtTemplateVars(
+  includeBundles: {'snackbar_sender'},
+  dependencies: [
+    "at_client_mobile: ^3.0.3",
+    "at_utils: ^3.0.0",
+    "path_provider: ^2.0.5",
+    "flutter_dotenv: ^5.0.2",
+    "at_app_flutter: ^5.0.0"
+  ],
+  kotlinVersion: Version.parse('1.5.32'),
+  minSdkVersion: '23',
+  flutterConfig: ["assets:", "  - .env"],
+);
+
+final AtAppTemplate snackbarSenderTemplate = AtAppTemplate(
   name: 'snackbar_sender',
   description: 'Send snackbars to a flutter web application publically.',
-  vars: AtTemplateVars(
-    includeBundles: {'snackbar_sender'},
-    dependencies: [
-      "at_client_mobile: ^3.0.3",
-      "at_utils: ^3.0.0",
-      "path_provider: ^2.0.5",
-      "flutter_dotenv: ^5.0.2",
-      "at_app_flutter: ^5.0.0"
-    ],
-    kotlinVersion: Version.parse('1.5.32'),
-    minSdkVersion: '23',
-    flutterConfig: ["assets:", "  - .env"],
-  ),
+  vars: _vars,
   overrideEnv: true,
   env: {"NAMESPACE": "fourballcorporate9", "API_KEY": "477b-876u-bcez-c42z-6a3d"},
-  bundles: [BaseTemplateBundle(), AndroidTemplateBundle(), IosTemplateBundle(), SnackbarSenderTemplateBundle()],
+  bundles: [
+    BaseTemplateBundle(),
+    AndroidTemplateBundle(),
+    IosTemplateBundle(),
+    SnackbarSenderTemplateBundle(),
+  ],
 );
