+++ packages/at_app/lib/src/bundles/samples/at_contacts_flutter/at_contacts_flutter_bundle.dart
@@ -40,5 +40,6 @@ final atContactsFlutterBundle = MasonBundle.fromJson(<String, dynamic>{
   "name": "at_contacts_flutter",
   "description": "A sample of how to use the at_contacts_flutter package.",
   "version": "0.1.0",
+  "environment": {"mason": "any"},
   "vars": {}
 });
