--- packages/at_app/lib/src/bundles/samples/at_contacts_flutter/at_contacts_flutter_template_bundle.dart
@@ -1,8 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-import 'package:at_app_create/at_app_create.dart';
-import 'at_contacts_flutter_bundle.dart';
-
-class AtContactsFlutterTemplateBundle extends AtTemplateBundle<AtTemplateVars> {
-  AtContactsFlutterTemplateBundle() : super(atContactsFlutterBundle);
-}
