+++ packages/at_app/lib/src/bundles/samples/at_onboarding_flutter/at_onboarding_flutter_bundle.dart
@@ -34,5 +34,6 @@ final atOnboardingFlutterBundle = MasonBundle.fromJson(<String, dynamic>{
   "name": "at_onboarding_flutter",
   "description": "A sample of how to use the at_onboarding_flutter package.",
   "version": "0.1.0",
+  "environment": {"mason": "any"},
   "vars": {}
 });
