--- packages/at_app/lib/src/bundles/samples/at_onboarding_flutter/at_onboarding_flutter_template_bundle.dart
@@ -1,8 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-import 'package:at_app_create/at_app_create.dart';
-import 'at_onboarding_flutter_bundle.dart';
-
-class AtOnboardingFlutterTemplateBundle extends AtTemplateBundle<AtTemplateVars> {
-  AtOnboardingFlutterTemplateBundle() : super(atOnboardingFlutterBundle);
-}
