+++ packages/at_app/lib/src/bundles/templates/app/app_bundle.dart
@@ -22,5 +22,6 @@ final appBundle = MasonBundle.fromJson(<String, dynamic>{
   "name": "app",
   "description": "The @platform skeleton app template.",
   "version": "0.1.0",
+  "environment": {"mason": "any"},
   "vars": {}
 });
