--- packages/at_app/lib/src/bundles/templates/app/app_template_bundle.dart
@@ -1,8 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-import 'package:at_app_create/at_app_create.dart';
-import 'app_bundle.dart';
-
-class AppTemplateBundle extends AtTemplateBundle<AtTemplateVars> {
-  AppTemplateBundle() : super(appBundle);
-}
